ca end

+++ lib/initializers.py
@@ -8,20 +8,20 @@ class Initializers:
     @classmethod
     def he_init(cls, shape:tuple):
         std_dev = np.sqrt(2.0 / shape[0])
-        return cls.rng.random(size=shape, dtype=np.float64) * std_dev
+        return cls.rng.random(size=shape, dtype=np.float32) * std_dev
     
 
     @classmethod
     def random_init(cls, shape:tuple):
-        return cls.rng.random(size=shape, dtype=np.float64)
+        return cls.rng.random(size=shape, dtype=np.float32)
     
 
     @classmethod
     def xavier_init(cls, shape:tuple):
         std_dev = np.sqrt(2.0 / (shape[0] + shape[1]))
-        return cls.rng.random(size=shape, dtype=np.float64) * std_dev
+        return cls.rng.random(size=shape, dtype=np.float32) * std_dev
 
 
     @staticmethod
     def constant_init(val, shape:tuple):
-        return np.full(shape=shape, fill_value=val, dtype=np.float64)
+        return np.full(shape=shape, fill_value=val, dtype=np.float32)
